fix(backend): fix disabling of S3
